requirements.txt
fastapi
uvicorn
pandas
requests
pyyaml
openpyxl>=3.1.0
python-multipart
PyGithub
